[MODULARIZE] - Refactor code in modules [DI] - Adding repository, usecases & viewmodel to DI
