+++ data/src/main/java/es/andres/bailen/data/di/DataModule.kt
@@ -1,5 +1,7 @@
 package es.andres.bailen.data.di
 
+import es.andres.bailen.data.repository.CharactersRepositoryImpl
+import es.andres.bailen.domain.repository.CharactersRepository
 import org.koin.core.module.Module
 import org.koin.dsl.module
 
@@ -9,7 +11,7 @@ val networkModule: Module = module {
 
 
 val repositoryModule: Module = module {
-
+    single <CharactersRepository>{ CharactersRepositoryImpl() }
 }
 
 val dataModules = listOf(networkModule, repositoryModule)
