+++ domain/src/main/java/es/andres/bailen/domain/repository/CharactersRepository.kt
@@ -1,0 +1,9 @@
+package es.andres.bailen.domain.repository
+
+import es.andres.bailen.domain.models.CharacterModel
+
+interface CharactersRepository {
+    fun getCharacters(): List<CharacterModel>
+
+    fun getCharacterDetail(characterId: String): CharacterModel?
+}
