+++ domain/src/main/java/es/andres/bailen/domain/usecases/CharacterListUseCase.kt
@@ -1,0 +1,10 @@
+package es.andres.bailen.domain.usecases
+
+import es.andres.bailen.domain.models.CharacterModel
+
+interface CharacterListUseCase {
+
+    fun getCharacterList(): List<CharacterModel>
+
+    fun getCharacterDetail(characterId: String): CharacterModel?
+}
